Script to Run Model Training
You can save all the feature subset csv's to the feature_subsets directory and run PostFeatureSelectionModelTraining to automatically evaluate all the models on these different feature subsets. If a model has already been run for a particular dataset, then the previously run model will be loaded in instead of retraining it. The data is added to results.csv in the results directory.